requirements.txt
alembic==1.11.1
apispec==6.3.0
blinker==1.6.2
cachelib==0.9.0
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-Caching==2.0.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Migrate==4.0.4
flask-smorest==0.42.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.1
passlib==1.7.4
Pillow==9.5.0
PyJWT==2.7.0
pypng==0.20220715.0
python-decouple==3.8
qrcode==7.4.2
six==1.16.0
SQLAlchemy==2.0.14
typing_extensions==4.5.0
webargs==8.2.0
Werkzeug==2.3.4
